코드 리팩토링 및 테스트 코드 추가
- 테스트 코드 : 통계 데이터

+++ src/main/java/com/example/passweb/entity/statistics/StatisticsEntity.java
@@ -20,9 +20,9 @@ public class StatisticsEntity {
 
     private LocalDateTime statisticsAt;
 
-    private Long allCount;
-    private Long attendedCount;
-    private Long cancelledCount;
+    private int allCount;
+    private int attendedCount;
+    private int cancelledCount;
